requirements.txt
channels==0.17.2
Django==1.10.1
factory-boy==2.7.0
git+https://github.com/jazzband/django-debug-toolbar.git#egg=django-debug-toolbar
pytest-django
